lab4/requirements.txt
fastapi
uvicorn
requests
tenacity
grpcio
grpcio-tools
hazelcast-python-client
kafka-python
python-json-logger